requirements.txt
google-generativeai>=0.8.0
pystray
Pillow
pyautogui
pynput
pyperclip
python-dotenv
pywin32